change name to parser to tokenizer
