+++ tokenizer/helper.go
@@ -1,4 +1,4 @@
-package parser
+package tokenizer
 
 import(
 	"fmt"
